Add #possible_moves for Pawn, Rook; pass Board
Adds possible_moves for Pawn and Rook, and Piece is passed Board.

possible_moves is a method in Piece, which iterates over the moveset of
a piece until another piece is blocking the path or if the location is
off the board. Rook's moveset is one square forward or back in col or
row. Pawn's possible_moves is overwritten to only show an advancement of
one square unless it has not moved, which adds the possibility for two
squares.

To accomodate checking whether a move is possible or not, Piece is now
passed the Board instance, and that change is reflected in each Piece
child class.